// src/app/shared/interface/profile/profile.interface.ts
export interface IProfileRequest{
    address: string;
    dateOfBirth: string;
    email: string;
    firstName: string;
    lastName: string;
    order: any;
    phoneNumber: string;
    pizza: string;
    role: string;
    price: string
}
export interface IProfileResponse{
    id: string | number;
    address: string;
    dateOfBirth: string;
    email: string;
    firstName: string;
    lastName: string;
    order: any;
    phoneNumber: string;
    pizza: string;
    role: string;
    price: string
}